requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Session==0.8.0
SQLAlchemy==2.0.46
mysql-connector-python==8.2.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.1.1
python-dateutil==2.8.2
